requirements.txt
mysql-connector-python>=8.0.0
pandas>=1.5.0
numpy>=1.20.0

